Fixed problem with pdflatex refusing to make the document -- problem was with one of the figures, which does not comply with PDF/A standard (may have to change it, now it is just commented out).
The style in the thesis shall not be numbering, but text (authors, year).
So I had to do some changes -- the simplest was to
turn all references from \cite to \citep (adds paretheses around the citation).

Thesis shall contain all relevant parts from ECC-POPC paper, now it is ready to be dissected.